Move original Admin into Core & rewrite dependencies
