+++ eventmesh-dashboard-core/src/test/java/org/apache/eventmesh/dashboard/core/EventMeshDashboardApplicationTest.java
@@ -1,0 +1,13 @@
+package org.apache.eventmesh.dashboard.core;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EventMeshDashboardApplicationTest {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
