+++ pom.xml
@@ -79,8 +79,8 @@
     <maven.compiler.source>8</maven.compiler.source>
     <maven.compiler.target>8</maven.compiler.target>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-
   </properties>
+
   <modules>
     <module>eventmesh-dashboard-console</module>
     <module>eventmesh-dashboard-observe</module>
@@ -88,6 +88,4 @@
     <module>eventmesh-dashboard-service</module>
     <module>eventmesh-dashboard-common</module>
   </modules>
-
-
 </project>
